moving order structure back into archive
